requirements.txt
absl-py==2.3.1
certifi==2025.7.9
charset-normalizer==3.4.2
clip @ git+https://github.com/openai/CLIP.git
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
et-xmlfile==2.0.0
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
ftfy==6.3.1
grpcio==1.71.0
hf-xet==1.1.7
huggingface-hub==0.34.4
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
markdown==3.8.2
markupsafe==2.1.1
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
numpy==1.24.4
open-clip-torch==3.1.0
openpyxl==3.1.5
packaging==25.0
pandas==1.4.2
pillow==11.3.0
protobuf==5.29.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
scikit-learn==1.1.1
scipy==1.12.0
seaborn==0.13.2
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.0
tensorboardx==2.6.4
threadpoolctl==3.6.0
timm==1.0.19
tokenizers==0.19.1
tqdm==4.67.1
transformers==4.40.0
triton==2.1.0
typing-extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
werkzeug==3.1.3
wheel==0.45.1
xlrd==2.0.2
zipp==3.23.0